Better OfferPage 2

+++ Explorers Haven/ViewModels/Booking/BookingViewModel.cs
@@ -7,7 +7,7 @@ namespace Explorers_Haven.ViewModels.Booking
         public int Id { get; set; }
 
         public decimal? PeopleCount { get; set; }
-        public int? YoungOldPeopleCount { get; set; }
+        public decimal? YoungOldPeopleCount { get; set; }
         public DateOnly? StartDate { get; set; }
         public decimal? Price { get; set; }
